How to detect anomaly using SNMP OIDs
This project aims to design and implement a system that is able to monitor the network using SNMP andidentify the specific possible attacks (DoS and port scan) using a cluster analysis. In the first task, theprogram discovers the topology of the network. After successful discovery phase, it will be able tomonitor the link utilization (network link-states) for a specified period of time, and then detect theanomaly, using k-means clustering scheme. These anomalies will be analyzed to recognize the attack.Moreover, this program also provides an advance feature, which is defined as optional task, as itexecutes online monitoring and detects the attacks using Davies-Bouldin Index as quality scoringmeasurement.